test: consolidate duplicate ExchangedFiatTests and stray Database test files
