+++ FlipcashCore/Tests/FlipcashCoreTests/PublicKey+TestSupport.swift
@@ -1,0 +1,5 @@
+import FlipcashCore
+
+extension PublicKey {
+    static let jeffy = try! PublicKey(base58: "54ggcQ23uen5b9QXMAns99MQNTKn7iyzq4wvCW6e8r25")
+}
