+++ FlipcashTests/Database/Database+ProfileTests.swift
@@ -1,5 +1,5 @@
 //
-//  DatabaseProfileTests.swift
+//  Database+ProfileTests.swift
 //  FlipcashTests
 //
